Extract WPCom authentication to a separate component
