+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/common/wpcomwebview/WPComWebViewAuthenticator.kt
@@ -1,0 +1,43 @@
+package com.woocommerce.android.ui.common.wpcomwebview
+
+import android.webkit.WebView
+import com.woocommerce.android.extensions.isNotNullOrEmpty
+import com.woocommerce.android.util.WooLog
+import org.wordpress.android.fluxc.store.AccountStore
+import java.io.UnsupportedEncodingException
+import java.net.URLEncoder
+import java.nio.charset.StandardCharsets
+import javax.inject.Inject
+
+private const val WPCOM_LOGIN_URL = "https://wordpress.com/wp-login.php"
+
+class WPComWebViewAuthenticator @Inject constructor(
+    private val accountStore: AccountStore
+) {
+    fun authenticateAndLoadUrl(webView: WebView, url: String) {
+        val postData = getAuthPostData(url)
+        webView.postUrl(WPCOM_LOGIN_URL, postData.toByteArray())
+    }
+
+    private fun getAuthPostData(redirectUrl: String): String {
+        val username = accountStore.account.userName.takeIf { it.isNotNullOrEmpty() } ?: return ""
+        val token = accountStore.accessToken.takeIf { it.isNotNullOrEmpty() } ?: return ""
+
+        val utf8 = StandardCharsets.UTF_8.name()
+        try {
+            var postData = String.format(
+                "log=%s&redirect_to=%s",
+                URLEncoder.encode(username, utf8),
+                URLEncoder.encode(redirectUrl, utf8)
+            )
+
+            // Add token authorization
+            postData += "&authorization=Bearer " + URLEncoder.encode(token, utf8)
+
+            return postData
+        } catch (e: UnsupportedEncodingException) {
+            WooLog.e(WooLog.T.UTILS, e)
+        }
+        return ""
+    }
+}
